refactor: Separation of the clients
Refs #FLU-213
