+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './controllers/index.js';
 export * from './services/index.js';
 export * from './types/index.js';
 export * from './prompts/index.js';
+export * from './clients/index.js';
