+++ src/services/index.ts
@@ -1,5 +1,5 @@
 export * from './mcp-gateway.service.js';
-export * from './fluora-mcp-client-sse.js';
-export * from './fluora-mcp-client-streamable.js';
+export * from '../clients/fluora-mcp-client-sse.js';
+export * from '../clients/fluora-mcp-client-streamable.js';
 export * from './fluora-api.service.js';
 export * from './blockchain-payment.service.js';
