+++ src/services/mcp-gateway.service.ts
@@ -1,13 +1,8 @@
-import { FluoraMcpClientSSE } from './fluora-mcp-client-sse.js';
-import { FluoraMcpClientStreamable } from './fluora-mcp-client-streamable.js';
-
-export interface FluoraMcpClient {
-  connect(mcpServerUrl: string): Promise<void>;
-  disconnect(): Promise<void>;
-  callTool(toolName: string, toolParams?: Record<string, unknown>): Promise<unknown>;
-  listTools(): Promise<unknown>;
-}
-
+import {
+  FluoraMcpClientSSE,
+  FluoraMcpClientStreamable,
+  FluoraMcpClient,
+} from '../clients/index.js';
 /**
  * Service for managing connections and interactions with MonetizedMCPServers
  * This is the core gateway that maintains compatibility with existing tools
